Applying bard to cicc
Firstly, NB there was a typo in the reduced_ham function that I have been using throughout this and the last project! in the [1,1] element, JK2/2 *(S-1) should
have been (JK/2)*S. I changed it but need to check earlier results.
I just hardcoded in the \pm basis ham from my paper anyway. Also need to check results against that.

I applied bardeen to the cicc model but note that
1) my kernel_well_super code fails when D is nonzero and S>1/2, ie the channel energies become spin dependent. need to fix this
2) since I currently use only one on-site spin ham to determine the change of basis, it is not clear whether the JK1 or JK2 should be used, so I just combine them. Also want to not have to do this