Generalize run_stt_ed.py to benchmark arbitrarily long Ne1 and Ne1_NFM2 systems
To get better agreement with Adrian, I try:
Using Jsd=2.0 (doesn’t work, but opens up wf norm decay, see next slide comments)
Using tdvp
Using bdim=750
Latter get a little but not a lot closer to Adrian’s results. Seem to converge to each other

1) When Jsd=2.0, default Super TST, the wf norm decays to 0.91 at t=12 (data/stt/demo/test_Jsd.out)
2)When Jsd=2.0 and dt=0.05, the wf decays to 0.94 (slurm/out/demoNe10.out)
3)When Jsd=2.0 and tdvp is used, the wf decays to 0.97 ( data/stt/demo/test_tdvp_wfdecay.out)
NB for all cases, at t=9 things are much better (0.999, 0.999, 1.000 respectively)
2) And 3) are very similar time costs so it is best to just to tdvp
Finally, for Jsd=0.5 and 1) it is already only 0.98 and with Jsd=0.5 and Ne=5 it is 0.9999. So we worry only when interactions are large AND Ne is large AND at long run times